fl-four-clients/requirements.txt
flwr==1.18.0
matplotlib==3.6.2
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.6.1
seaborn==0.13.2
torch==2.5.1
torchmetrics==1.7.0
